Supply information for emptiness
Allows to add more informations for emptiness algorithms. This
additionnal information can be used to pass the two automata of a
product (if the emptiness needs to check data into these
automata). For instance, this can be used to get the projection of a
state into the formula automaton in order to get its formula.

* src/tgbaalgos/Makefile.am: Add emptiness supply information files
* src/tgbaalgos/emptiness.cc: add all new algorithms
* src/tgbaalgos/emptiness.hh: add default specifier
* src/tgbaalgos/emptiness_specifier.cc: Implementation of the
  specifier
* src/tgbaalgos/emptiness_specifier.hh: Provide information that is
  needed when performing the emptiness