remove_x: generalize to siPSL following Dax et al. ATVA'09
DO NOT MERGE THIS AS IS

This implements the κ (Lemma 4) and τ (Lemma 9) transformations from
the Dax et al. paper at ATVA'09.  Unfortunately testing this revealed
two issues (one in each of these lemmas) that I'm not sure how to fix.

Here is a counterexample for Lemma 4:

  L#(a && (a;a)) = ∅   but  L#(κ(a && (a;a))) = a[+]

Here is a counterexample for Lemma 9:

  L({a}<>->b) = L(a & b)
  L(τ({a}<>->b)) = L({a[+]}<>->b) = L(b M a)

and clearly L#(a & b) != L#(b M a).

I've sent an email to the paper's authors.

* src/ltlvisit/remove_x.cc: Implement the rewriting.
* src/ltltest/remove_x.test: Add a commented test to show the above two
issues.