Fix GITPATCH computation
Since the introduction of the syntcomp2018-submission tag (on a
separate branch), that computation of GITPATCH was based on this
tag because it was the last one (even if it is not on next).
We did not see it on the Debian packages, because they clone
only one branch of the repository, but the RPM packages had a
different GITPATCH.

* configure.ac (GITPATCH): Compute the number of commits since the
last "spot-*" tag.