Define new state label functions + their defaults
This API enables access to "labels" in LTSmin/PINS parlance, which are
atomic propositions of the formula in our setting.

We provide defaults that should always work (empty set of labels) to
retro-adapt older targets not presenting these functions.

Defaults are defined in the same file, passing lambdas makes compiler
complain about creating references to temporaries, which is legitimate.